Clean code for Employee duplicated in Security
