+++ Api/Employee/Application/Assembler/EmployeeAssembler.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
 using InkaPharmacy.Api.Employees.Application.Dto;
+using InkaPharmacy.Api.Employees.Domain.Entity;
+using System.Collections.Generic;
 
 namespace InkaPharmacy.Api.Employees.Application.Assembler
 {
-    using InkaPharmacy.Api.Employees.Domain.Entity;
+
 
     public class EmployeeAssembler
     {
@@ -14,9 +16,19 @@ namespace InkaPharmacy.Api.Employees.Application.Assembler
             _mapper = mapper;
         }
 
-        public Employee toEntity(EmployeeDto EmployeeLoginDto)
+        public Employee toEntity(EmployeeDto EmployeeDto)
         {
-            return _mapper.Map<Employee>(EmployeeLoginDto);
+            return _mapper.Map<Employee>(EmployeeDto);
+        }
+
+        public EmployeeDto toDto(Employee employee)
+        {
+            return _mapper.Map<EmployeeDto>(employee);
+        }
+
+        public List<EmployeeDto> toDtoList(List<Employee> employeeList)
+        {
+            return _mapper.Map<List<Employee>, List<EmployeeDto>>((List<Employee>)employeeList);
         }
     }
