+++ Api/Security/Infrastructure/Persistence/NHibernate/Specification/LoginBySpecification.cs
@@ -1,28 +1,24 @@
 ﻿using InkaPharmacy.Api.Common.Domain.Specification;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using InkaPharmacy.Api.Employees.Domain.Entity;
+using System.Linq.Expressions;
 
 namespace InkaPharmacy.Api.Security.Infrastructure.Persistence.NHibernate.Specification
 {
-    using InkaPharmacy.Api.Employees.Domain.Entity;
-    using System.Linq.Expressions;
-
     public class LoginBySpecification : Specification<Employee>
     {
-        private readonly string _usu;
-        private readonly string _clave;
+        private readonly string _username;
+        private readonly string _password;
 
-        public LoginBySpecification(string usu, string clave)
+        public LoginBySpecification(string username, string password)
         {
-            _usu = usu;
-            _clave = clave;
+            _username = username;
+            _password = password;
         }
 
         public override Expression<Func<Employee, bool>> ToExpression()
         {
-            return employee => (employee.Password == _clave && employee.Username == _usu);
+            return employee => (employee.Password == _password && employee.Username == _username);
         }
 
     }
